pyproject.toml
[project]
name = "z-ai-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "pymongo>=4.15.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "sniffio>=1.3.0",
    "zai-sdk>=0.2.0",
]
